+++ docs/CHANGE_LOG.md
@@ -1,5 +1,27 @@
 # Observe.js - Change Log:
 
+## Version 0.3.0:
+
+#### **New Features:**
+- Added "Pause Timeout On Change" support, which will force the watch to wait a specific number of milliseconds before detecting changes again when a change is detected.
+
+#### **Binding Options / Function Options:**
+- Added a new binding/option called "pauseTimeoutOnChange", which states the milliseconds to wait for new changes to be detected when a change is detected (defaults to 0, which is off).
+
+#### **Public Functions:**
+- Added new public function "cancelWatches()", which is used to cancel all the watches currently running, or paused.
+- Added new public function "getWatches()", which is used to return all the watches currently running, or paused.
+- Added new public function "pauseWatch()", which is used to pause a running watch for a specific number of milliseconds.
+
+#### **General Improvements:**
+- When the page is unloaded, all active watches are now cancelled.
+
+#### **Documentation:**
+- Fixed some of the documentation spelling/grammar mistakes.
+
+<br>
+
+
 ## Version 0.2.0:
 
 #### **New Features:**